feat: add feature like/dislike, delete card

+++ src/utils/api.js
@@ -1,4 +1,5 @@
 import handleOriginalResponse from './utils.js';
+
 class Api {
   constructor(options) {
     this.baseUrl = options.baseUrl;
@@ -47,24 +48,34 @@ class Api {
     return fetch(`${this.baseUrl}/cards/${cardId}`, {
       method: 'DELETE',
       headers: this.headers,
-    });
-  }
-
-  addLike(cardId) {
-    return fetch(`${this.baseUrl}/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this.headers,
     })
       .then(handleOriginalResponse);
+    ;
   }
 
-  deleteLike(cardId) {
-    return fetch(`${this.baseUrl}/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this.headers,
-    })
-      .then(handleOriginalResponse);
+  changeLikeCardStatus(cardId, setLike) {
+    if (setLike) {
+      return fetch(`${this.baseUrl}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this.headers,
+      })
+        .then(handleOriginalResponse);
+    } else {
+      return fetch(`${this.baseUrl}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this.headers,
+      })
+        .then(handleOriginalResponse);
+    }
   }
+
+  // addLike(cardId) {
+
+  // }
+
+  // deleteLike(cardId) {
+
+  // }
 
   updateAvatar(avatarLink) {
     return fetch(`${this.baseUrl}/users/me/avatar`, {
